Гуи и контерней Аккумулятора
Аккумулятор имеет своё ГУИ
Аккумулятор имеет свой контейнер
Батарейка отображает правильное значение энергии (1 повреждения = 0 заряда, макс - полный заряд)
Название энергии вынесено в константы для удобства
Исправлена ошибка с энергией - энергия была short, теперь int

+++ src/main/java/ru/nord/common/tiles/abstracts/TileAbstractEnergyGenerator.java
@@ -154,7 +154,7 @@ public abstract class TileAbstractEnergyGenerator extends TileAbstractEnergyBloc
         }
 
 
-        this.setEnergy(compound.getShort("energy"));
+        this.setEnergy(compound.getInteger("energy"));
         burnTime = compound.getShort("burnTime");
         if (compound.hasKey("CustomName", Constants.NBT.TAG_STRING)) {
             this.machineCustomName = compound.getString("CustomName");
@@ -165,7 +165,7 @@ public abstract class TileAbstractEnergyGenerator extends TileAbstractEnergyBloc
     public void writeToNBT(NBTTagCompound compound) {
         super.writeToNBT(compound);
 
-        compound.setShort("energy", (short) this.getEnergy());
+        compound.setInteger("energy", (short) this.getEnergy());
         compound.setShort("burnTime", (short) burnTime);
         NBTTagList itemList = new NBTTagList();
         for (int i = 0; i < this.inventory.length; ++i) {
